ShoppingKitAWS/requirements.txt
Django==2.1.1
django-appconf==1.0.3
django-contrib-comments==1.9.1
django-imagekit==4.0.2
django-vote==2.1.7
pilkit==2.0
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
six==1.12.0
